+++ cmd/web.go
@@ -7,9 +7,9 @@ import (
 // TODO: Mudar pra cada perfil quando finalizar esse perfil
 // ProfileCmd comando para configurar o perfil de desenvolvedor web
 var ProfileCmd = &cobra.Command{
-	Use:   "Profiles",
-	Short: "Configura o perfil de desenvolvedor",
-	Long:  "Permite selecionar ferramentas específicas para desenvolvimento.",
+	Use:   "profile-web",
+	Short: "Configura o perfil de desenvolvedor web",
+	Long:  "Permite selecionar ferramentas específicas para desenvolvimento web.",
 	Run: func(cmd *cobra.Command, args []string) {
 		// Chamando a função para configurar o perfil web
 		config.ConfigureWebProfile()